Memory allocations are now all tracked with size_t instead of plain ol' int.
TNT2 MAP30 does things to my port. This is step one to bringing it back to some measure of sensible.

+++ src/z_zone.h
@@ -53,7 +53,7 @@ enum
         
 
 void	Z_Init (void);
-void*	Z_Malloc (int size, int tag, void *ptr);
+void*	Z_Malloc( size_t size, int tag, void *ptr );
 void    Z_Free (void *ptr);
 void    Z_FreeTags (int lowtag, int hightag);
 void    Z_DumpHeap (int lowtag, int hightag);
@@ -61,8 +61,8 @@ void    Z_FileDumpHeap (FILE *f);
 void    Z_CheckHeap (void);
 boolean Z_ChangeTag2 (void *ptr, int tag, const char *file, int line);
 void    Z_ChangeUser(void *ptr, void **user);
-int     Z_FreeMemory (void);
-unsigned int Z_ZoneSize(void);
+size_t	Z_FreeMemory (void);
+size_t	Z_ZoneSize(void);
 
 //
 // This is used to get the local FILE:LINE info from CPP
